moved zcl reload testing to zcl-loader-reload.test.js, added duplicate zcl.json and general.xml files for this test
