+++ test/zcl-loader.test.js
@@ -189,50 +189,6 @@ test(
           `Found Non Unique Enum in Silabs XML: ${c.NAME} ${c.TYPE} ${c.PACKAGE_REF}`
         )
       })
-
-      // Test Reloading of the zcl packages when an xml file changes
-      // (Simulating a gsdk upgrade from one branch to another which can change xml files)
-      // Step 1: Modify one of the xml files
-      let xmlFilePath = env.builtinSilabsZclGeneralXmlFile()
-      let originalString =
-        '<attribute side="server" code="0x0001" define="APPLICATION_VERSION" type="INT8U" min="0x00" max="0xFF" writable="false" default="0x00" optional="true">application version</attribute>'
-      let editString =
-        '<attribute side="server" code="0x0001" define="APPLICATION_VERSION" type="INT8U" min="0x00" max="0xFF" writable="false" default="0x01" optional="true">application version</attribute>'
-      let generalXmlFileOriginalContent = fs.readFileSync(xmlFilePath, 'utf8')
-      let generalXmlFileUpdatedContent = generalXmlFileOriginalContent.replace(
-        originalString,
-        editString
-      )
-      fs.writeFileSync(xmlFilePath, generalXmlFileUpdatedContent, 'utf8')
-
-      // Step 2: This will cause the all zcl packages from top level to all its
-      // children to be reloaded. Check for 2 top level packages with same path now.
-      // Count packages belonging to each top level package to make sure all packages
-      // are reloaded
-      ctx = await zclLoader.loadZcl(db, env.builtinSilabsZclMetafile())
-      let newPackageId = ctx.packageId
-      // Making sure the top level packages do not have the same packageId
-      expect(packageId).not.toEqual(newPackageId)
-      let oldPackages = await dbApi.dbAll(
-        db,
-        `SELECT * FROM PACKAGE WHERE PARENT_PACKAGE_REF = ${packageId}`
-      )
-      let newPackages = await dbApi.dbAll(
-        db,
-        `SELECT * FROM PACKAGE WHERE PARENT_PACKAGE_REF = ${newPackageId}`
-      )
-      // Making sure all packages are loaded again
-      expect(oldPackages.length).toEqual(newPackages.length)
-
-      let topLevelZclPackages = await dbApi.dbAll(
-        db,
-        `SELECT * FROM PACKAGE WHERE TYPE = '${dbEnum.packageType.zclProperties}' ORDER BY PACKAGE_ID`
-      )
-      expect(topLevelZclPackages[0].IS_IN_SYNC).toEqual(0)
-      expect(topLevelZclPackages[1].IS_IN_SYNC).toEqual(1)
-
-      // Step 3: Revert the xml file change in step 1.
-      fs.writeFileSync(xmlFilePath, generalXmlFileOriginalContent, 'utf8')
     } finally {
       await dbApi.closeDatabase(db)
     }
