+++ zcl-builtin/silabs/zcl-special.json
@@ -1,0 +1,71 @@
+{
+  "version": 1,
+  "description": "ZigbeePro test data",
+  "category": "zigbee",
+  "xmlRoot": ".",
+  "xmlFile": [
+    "chip.xml",
+    "types.xml",
+    "general-special.xml",
+    "ha.xml",
+    "ha-devices.xml",
+    "cba.xml",
+    "cba-devices.xml",
+    "ota.xml",
+    "ami.xml",
+    "ami-devices.xml",
+    "zll.xml",
+    "zll-devices.xml",
+    "ta.xml",
+    "ta-devices.xml",
+    "hc.xml",
+    "hc-devices.xml",
+    "green-power.xml",
+    "green-power-devices.xml",
+    "silabs.xml",
+    "lo-devices.xml",
+    "wwah-silabs.xml",
+    "wwah-silabs-devices.xml",
+    "demo.xml",
+    "sample-extensions.xml"
+  ],
+  "zclSchema": "../shared/schema/zcl.xsd",
+  "supportCustomZclDevice": true,
+  "manufacturersXml": "../shared/manufacturers.xml",
+  "profilesXml": "../shared/profiles.xml",
+  "zclValidation": "./schema/zcl-validation.js",
+  "options": {
+    "text": {
+      "defaultResponsePolicy": ["Always", "Conditional", "Never"]
+    },
+    "bool": ["commandDiscovery"]
+  },
+  "defaults": {
+    "text": {
+      "manufacturerCodes": "0x1049",
+      "defaultResponsePolicy": "always"
+    },
+    "bool": {
+      "commandDiscovery": true
+    }
+  },
+  "featureFlags": {
+    "validationTimersFlags": {
+      "zigbeeEnforceCommonCluster": true
+    }
+  },
+  "ZCLDataTypes": ["ARRAY", "BITMAP", "ENUM", "NUMBER", "STRING", "STRUCT"],
+  "uiOptions": {
+    "showProfileId": true
+  },
+  "zcl": {
+    "multiProtocol": {
+      "category": ["zigbee", "matter"],
+      "type": "boolean",
+      "configurability": "hidden",
+      "globalDefault": false,
+      "label": "List of clusters and attributes associated across protocols",
+      "defaults": "multi-protocol.json"
+    }
+  }
+}
